finish merger
